SELinux: use define for number of bits in the mnt flags mask
We had this random hard coded value of '8' in the code (I put it there)
for the number of bits to check for mount options.  This is stupid.  Instead
use the #define we already have which tells us the number of mount
options.

Signed-off-by: Eric Paris <eparis@redhat.com>
Change-Id: I63d2fba4f650a08de875a0b6aaef079da2242059
(cherry picked from commit 761dcd99c94e63d6bb4049e8dcb2a75b5e8fa1f8)